refactor unify1 and match1 - stage 3
